fix: update raster-shared dependency to version 1.9.0 and rename fileUtils to fileUtil
