+++ package.json
@@ -50,7 +50,7 @@
     "@map-colonies/mc-model-types": "^17.15.1",
     "@map-colonies/mc-priority-queue": "^8.1.1",
     "@map-colonies/mc-utils": "^3.2.0",
-    "@map-colonies/raster-shared": "file:../../raster-shared/map-colonies-raster-shared-1.8.0.tgz",
+    "@map-colonies/raster-shared": "^1.9.0",
     "@map-colonies/read-pkg": "0.0.1",
     "@map-colonies/telemetry": "^7.0.1",
     "@map-colonies/types": "^1.4.0",
